Part1 - Line is drawn perfectly fine!

+++ ImageDisplay_C++/ImageDisplay_C++/Image.h
@@ -48,7 +48,7 @@ public:
 	MyImage & operator= (const MyImage & otherImage);
 
 	// Reader & Writer functions
-	void    copy(const MyImage &inImage, int scale, float alias);
+	void    copy(const MyImage &inImage, float scale, float alias);
 	void	setWidth( const int w)  { Width = w; }; 
 	void	setHeight(const int h) { Height = h; }; 
 	void	setImageData( const char *img ) { Data = (char *)img; };
